Cover direct build graph file read effects
