+++ tests/integration/cli_build.rs
@@ -12,6 +12,8 @@ mod build_graph_json;
 mod diagnostics;
 #[path = "cli_build/direct_build_graph_execution.rs"]
 mod direct_build_graph_execution;
+#[path = "cli_build/direct_build_graph_host_effects.rs"]
+mod direct_build_graph_host_effects;
 #[path = "cli_build/emit_direct.rs"]
 mod emit_direct;
 #[path = "cli_build/emit_direct_host_effects.rs"]
